Converted symbols to use BigInteger and made the code smaller

+++ Codex/src/main/java/org/operatorfoundation/codex/symbols/Binary.kt
@@ -11,16 +11,16 @@ class Binary : Symbol {
 
     override fun decode(encodedValue: ByteArray): BigInteger {
         return when (encodedValue.decodeToString()) {
-            "0" -> 0
-            "1" -> 1
+            "0" -> 0.toBigInteger()
+            "1" -> 1.toBigInteger()
             else -> throw IllegalArgumentException("Binary, bad value $encodedValue")
         }
     }
 
     override fun encode(numericValue: BigInteger): ByteArray {
         return when (numericValue) {
-            0 -> "0".toByteArray()
-            1 -> "1".toByteArray()
+            0.toBigInteger() -> "0".toByteArray()
+            1.toBigInteger() -> "1".toByteArray()
             else -> throw IllegalArgumentException("Unknown value $numericValue for Binary")
         }
     }
