+++ Codex/src/main/java/org/operatorfoundation/codex/symbols/CallAny.kt
@@ -1,15 +1,23 @@
 package org.operatorfoundation.codex.symbols
 
+import java.math.BigInteger
+
 class CallAny : Symbol {
     companion object {
         private val charToValue = mapOf(
-            "A" to 0, "B" to 1, "C" to 2, "D" to 3, "E" to 4, "F" to 5,
-            "G" to 6, "H" to 7, "I" to 8, "J" to 9, "K" to 10, "L" to 11,
-            "M" to 12, "N" to 13, "O" to 14, "P" to 15, "Q" to 16, "R" to 17,
-            "S" to 18, "T" to 19, "U" to 20, "V" to 21, "W" to 22, "X" to 23,
-            "Y" to 24, "Z" to 25, "0" to 26, "1" to 27, "2" to 28, "3" to 29,
-            "4" to 30, "5" to 31, "6" to 32, "7" to 33, "8" to 34, "9" to 35,
-            " " to 36
+            "A" to BigInteger.valueOf(0), "B" to BigInteger.valueOf(1), "C" to BigInteger.valueOf(2),
+            "D" to BigInteger.valueOf(3), "E" to BigInteger.valueOf(4), "F" to BigInteger.valueOf(5),
+            "G" to BigInteger.valueOf(6), "H" to BigInteger.valueOf(7), "I" to BigInteger.valueOf(8),
+            "J" to BigInteger.valueOf(9), "K" to BigInteger.valueOf(10), "L" to BigInteger.valueOf(11),
+            "M" to BigInteger.valueOf(12), "N" to BigInteger.valueOf(13), "O" to BigInteger.valueOf(14),
+            "P" to BigInteger.valueOf(15), "Q" to BigInteger.valueOf(16), "R" to BigInteger.valueOf(17),
+            "S" to BigInteger.valueOf(18), "T" to BigInteger.valueOf(19), "U" to BigInteger.valueOf(20),
+            "V" to BigInteger.valueOf(21), "W" to BigInteger.valueOf(22), "X" to BigInteger.valueOf(23),
+            "Y" to BigInteger.valueOf(24), "Z" to BigInteger.valueOf(25), "0" to BigInteger.valueOf(26),
+            "1" to BigInteger.valueOf(27), "2" to BigInteger.valueOf(28), "3" to BigInteger.valueOf(29),
+            "4" to BigInteger.valueOf(30), "5" to BigInteger.valueOf(31), "6" to BigInteger.valueOf(32),
+            "7" to BigInteger.valueOf(33), "8" to BigInteger.valueOf(34), "9" to BigInteger.valueOf(35),
+            " " to BigInteger.valueOf(36)
         )
 
         private val valueToChar = charToValue.map { (k, v) -> v to k }.toMap()
@@ -19,12 +27,12 @@ class CallAny : Symbol {
 
     override fun toString(): String = "CallAny"
 
-    override fun decode(encodedValue: ByteArray): Int {
+    override fun decode(encodedValue: ByteArray): BigInteger {
         return charToValue[encodedValue.toString()]
             ?: throw IllegalArgumentException("CallAny, bad value: $encodedValue")
     }
 
-    override fun encode(numericValue: Int): ByteArray {
+    override fun encode(numericValue: BigInteger): ByteArray {
         return (valueToChar[numericValue]
             ?: throw IllegalArgumentException("Unknown value $numericValue for CallAny"))
             .toByteArray()
